refactor: move types into types
