+++ src/types/index.ts
@@ -10,6 +10,7 @@ export * from './login'
 export * from './myinfo_hash'
 export * from './response'
 export * from './sms_count'
+export * from './sns'
 export * from './submission'
 export * from './token'
 export * from './user'
