+++ src/types/sns.ts
